requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
defusedxml==0.7.1
distro==1.9.0
fastapi==0.95.1
frozenlist==1.3.3
h11==0.16.0
httpcore==1.0.9
httptools==0.5.0
httpx==0.28.1
idna==3.4
jira==3.5.0
jiter==0.9.0
multidict==6.0.4
numpy==2.2.5
oauthlib==3.2.2
openai==1.76.0
packaging==25.0
pandas==2.2.3
pydantic==1.10.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
redis==4.5.5
regex==2023.5.5
requests==2.30.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.0
starlette==0.26.1
tiktoken==0.4.0
tqdm==4.65.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.0.2
uvicorn==0.22.0
yarl==1.9.2
langchain==0.2.1
langchain-community==0.2.1